Move inbox comment and mention folders
